Move twoslash tests over to a new file
